// src/store-modules/list/lists-store-module-mutation-type.ts
enum LIST_STORE_MODULE_MUTATION_TYPE {
    LIST_FOCUSED_SET = "listFocusedSet",
    LIST_FOCUSED_SET_DOCUMENTS = "listFocusedSetDocuments",
    LIST_FOCUSED_SET_INDEX = "listFocusedSetIndex",
    LIST_SET_ALL = 'listSetAll',
    LIST_SET_MANY = 'listSetMany',
    LIST_SET_ONE = 'listSetOne',
    LIST_UNSET_ONE = 'listUnsetOne'
}

export default LIST_STORE_MODULE_MUTATION_TYPE;